+++ bluefin/CHANGELOG.md
@@ -1,3 +1,5 @@
+* Add `Bluefin.HandleReader.asksHandle`
+
 * Restrict type parameter of `PrimStateEff` to `Effects`
 
   This is technically a breaking change, but we did not enforce a